+++ user-application/src/main/java/io/github/dronesecurity/userapplication/events/DataRead.java
@@ -13,7 +13,7 @@ import java.util.Map;
 public class DataRead implements Event {
     private final double proximity;
     private final Map<String, Double> accelerometerData;
-    private final double cameraData;
+    private final long cameraData;
 
     /**
      * Instantiates the Data Read event.
@@ -22,7 +22,7 @@ public class DataRead implements Event {
      * @param accelerometer data read by the accelerometer
      * @param camera data read by the camera
      */
-    public DataRead(final double proximity, final Map<String, Double> accelerometer, final double camera) {
+    public DataRead(final double proximity, final Map<String, Double> accelerometer, final long camera) {
         this.proximity = proximity;
         this.accelerometerData = Map.copyOf(accelerometer);
         this.cameraData = camera;
@@ -51,7 +51,7 @@ public class DataRead implements Event {
      *
      * @return value read by the camera
      */
-    public double getCameraData() {
+    public long getCameraData() {
         return this.cameraData;
     }
 }
